backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.1
pydantic==2.10.0
pydantic-settings==2.6.1

# HTTP Client
httpx==0.27.2

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0
greenlet==3.1.1

# CORS
python-multipart==0.0.12

# Environment Variables
python-dotenv==1.0.1

# OpenTelemetry
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-instrumentation-httpx==0.49b2
opentelemetry-instrumentation-sqlalchemy==0.49b2
opentelemetry-exporter-otlp-proto-http==1.28.2

# Date/Time
python-dateutil==2.8.2

